chore(base): Move `Room::is_favourite` and `::is_low_priority` into new `tags` module.
This patch moves the  Room::is_favourite` and `::is_low_priority`, along
with the `RoomNotableTags` into the new `tags` module. This patch also
moves the tests in this new module.

The idea is to group API by “theme” to get smaller modules and more
organised code.
